platform/chatbot/pyproject.toml
[tool.poetry]
name = "faciliauto-chatbot"
version = "0.1.0"
description = "WhatsApp Chatbot for FacilIAuto - Intelligent vehicle recommendation assistant"
authors = ["FacilIAuto Team"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = "^2.5.0"
pydantic-ai = "^0.0.13"
langgraph = "^0.0.20"
langchain = "^0.1.0"
redis = "^5.0.1"
duckdb = "^0.9.2"
celery = "^5.3.4"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
prometheus-client = "^0.19.0"
cryptography = "^42.0.0"
pyjwt = "^2.8.0"
spacy = "^3.7.2"
transformers = "^4.36.0"
torch = "^2.1.0"
sentence-transformers = "^2.3.1"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.1.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
httpx-mock = "^0.10.0"
faker = "^22.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "duckdb.*",
    "spacy.*",
    "transformers.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing --cov-report=html"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*", "**/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
